// templates/gong/types.ts
export interface FilterFields {
    fromDateTime?: string | undefined;
    toDateTime?: string | undefined;
    workspaceId?: string | undefined;
    callIds?: string[] | undefined;
}
// ...
interface Sentence {
    start: number;
    end: number;
    text: string;
}

interface Transcript {
    speakerId: string;
    topic: string | null;
    sentences: Sentence[];
}

export interface GongCallTranscript {
    callId: string;
    transcript: Transcript[];
}
// ...
export interface AxiosError<E = any> {
    message: string;
    name: string;
    code: string;
    response?: {
        status: number;
        statusText: string;
        data: E;
    };
}

export interface GongError {
    requestId: string;
    errors: string[];
}